comprehensive_compliance_check.py
#!/usr/bin/env python3
"""
Comprehensive Compliance Check for mlTrainer
"""

import os
import sys
import importlib
from pathlib import Path
from typing import Dict, List, Any
from dotenv import load_dotenv

load_dotenv()


def check_environment():
    """Check environment setup"""
    print("\n🔍 Checking Environment# Production code implemented")

    # Check .env file
    env_file = Path(".env")
    if env_file.exists():
        print("✅ .env file exists")
    else:
        print("❌ .env file missing")
        return False

    # Check required environment variables
    required_vars = ["POLYGON_API_KEY", "FRED_API_KEY"]
    missing_vars = []

    for var in required_vars:
        if os.getenv(var):
            print(f"✅ {var} is set")
        else:
            print(f"❌ {var} is missing")
            missing_vars.append(var)

    if missing_vars:
        print(f"⚠️  Missing environment variables: {', '.join(missing_vars)}")
        return False

    return True


def check_dependencies():
    """Check required dependencies"""
    print("\n🔍 Checking Dependencies# Production code implemented")

    required_packages = [
        "numpy",
        "pandas",
        "scikit-learn",
        "matplotlib",
        "seaborn",
        "requests",
        "yfinance",
        "ta",
        "statsmodels",
        "prophet",
    ]

    missing_packages = []

    for package in required_packages:
        try:
            importlib.import_module(package)
            print(f"✅ {package} is installed")
        except ImportError:
            print(f"❌ {package} is missing")
            missing_packages.append(package)

    if missing_packages:
        print(f"⚠️  Missing packages: {', '.join(missing_packages)}")
        return False

    return True


def check_custom_modules():
    """Check custom module imports"""
    print("\n🔍 Checking Custom Modules# Production code implemented")

    custom_modules = [
        "custom.time_series",
        "custom.rl",
        "custom.meta_learning",
        "custom.ensemble",
        "custom.regime_detection",
        "custom.risk_management",
        "custom.financial_models",
        "custom.momentum",
        "custom.systems",
        "custom.detectors",
        "custom.adversarial",
        "custom.fractal",
        "custom.automl",
        "custom.nonlinear",
        "custom.position_sizing",
        "custom.risk",
        "custom.volatility",
        "custom.complexity",
        "custom.stress",
        "custom.microstructure",
        "custom.optimization",
        "custom.macro",
        "custom.pairs",
        "custom.adaptive",
        "custom.elliott_wave",
        "custom.binomial",
        "custom.interest_rate",
        "custom.alternative_data",
        "custom.regime_ensemble",
    ]

    failed_modules = []

    for module in custom_modules:
        try:
            importlib.import_module(module)
            print(f"✅ {module} imports successfully")
        except ImportError as e:
            print(f"❌ {module} import failed: {e}")
            failed_modules.append(module)

    if failed_modules:
        print(f"⚠️  Failed modules: {', '.join(failed_modules)}")
        return False

    return True


def check_model_implementations():
    """Check model implementations"""
    print("\n🔍 Checking Model Implementations# Production code implemented")

    # production a few key models
    test_models = [
        ("custom.time_series", "RollingMeanReversion"),
        ("custom.rl", "RegimeAwareDQN"),
        ("custom.meta_learning", "MetaLearnerStrategySelector"),
        ("custom.ensemble", "EnsembleVoting"),
        ("custom.regime_detection", "RollingZScoreRegimeScorer"),
        ("custom.risk_management", "DynamicRiskParityModel"),
        ("custom.financial_models", "BlackScholes"),
        ("custom.momentum", "CCIEnsemble"),
        ("custom.systems", "MomentumBreakout"),
        ("custom.detectors", "TrendReversal"),
    ]

    failed_models = []

    for module_name, class_name in test_models:
        try:
            module = importlib.import_module(module_name)
            model_class = getattr(module, class_name)

            # production instantiation
            instance = model_class()
            print(f"✅ {module_name}.{class_name} instantiated successfully")

        except Exception as e:
            print(f"❌ {module_name}.{class_name} failed: {e}")
            failed_models.append(f"{module_name}.{class_name}")

    if failed_models:
        print(f"⚠️  Failed models: {', '.join(failed_models)}")
        return False

    return True


def check_configuration():
    """Check configuration files"""
    print("\n🔍 Checking Configuration# Production code implemented")

    config_files = [
        "config/models_config.py",
        "config/api_config.py",
        "config/secrets_manager.py",
        "config/compliance_enforcer.py",
    ]

    for config_file in config_files:
        if Path(config_file).exists():
            print(f"✅ {config_file} exists")
        else:
            print(f"❌ {config_file} missing")
            return False

    return True


def check_compliance_rules():
    """Check compliance rules"""
    print("\n🔍 Checking Compliance Rules# Production code implemented")

    try:
        # production compliance enforcer import
        from config.compliance_enforcer import ComplianceEnforcer

        print("✅ ComplianceEnforcer imports successfully")

        # production basic compliance check
        ce = ComplianceEnforcer()
        print("✅ ComplianceEnforcer instantiated successfully")

        return True

    except Exception as e:
        print(f"❌ Compliance check failed: {e}")
        return False


def main():
    """Run comprehensive compliance check"""
    print("🔒 mlTrainer Comprehensive Compliance Check")
    print(("=" * 50))

    checks = [
        ("Environment Setup", check_environment),
        ("Dependencies", check_dependencies),
        ("Custom Modules", check_custom_modules),
        ("Model Implementations", check_model_implementations),
        ("Configuration", check_configuration),
        ("Compliance Rules", check_compliance_rules),
    ]

    results = []

    for check_name, check_func in checks:
        try:
            result = check_func()
            results.append((check_name, result))
        except Exception as e:
            print(f"❌ {check_name} check failed with error: {e}")
            results.append((check_name, False))

    # Summary
    print(("\n" + "=" * 50))
    print("📊 COMPLIANCE CHECK SUMMARY")
    print(("=" * 50))

    passed = 0
    total = len(results)

    for check_name, result in results:
        status = "✅ PASS" if result else "❌ FAIL"
        print(f"{status} {check_name}")
        if result:
            passed += 1

    print(f"\nOverall: {passed}/{total} checks passed")

    if passed == total:
        print("🎉 All compliance checks passed!")
        return True
    else:
        print("⚠️  Some compliance checks failed. Please address the issues above.")
        return False


if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)
