requirements.txt
# mlTrainer Standard Requirements
# =================================
# Standard dependencies for most use cases
# For full 125+ models, use requirements_comprehensive.txt
# For minimal setup, use requirements_minimal.txt
# Last Updated: July 2025

# Core Dependencies
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
joblib==1.3.2

# Machine Learning Models
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2
optuna==3.3.0  # Hyperparameter tuning

# Financial Data & Analysis
# ------------------------
yfinance==0.2.28
fredapi==0.5.1
polygon-api-client==1.12.0
ccxt==4.0.0  # Crypto exchanges
alpaca-trade-api==3.0.2
pandas-ta==0.3.14b0
ta-lib==0.4.28  # Requires system TA-Lib
pypfopt==1.5.5  # Portfolio optimization
arch==6.1.0  # Volatility modeling
backtrader==1.9.78.123
py-vollib==1.0.1  # Options pricing (Black-Scholes Greeks)
quantlib==1.31  # Quantitative finance library

# Web Framework & API
# -------------------
streamlit==1.28.0
streamlit-plotly-events==0.0.6
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.4.0
aiohttp==3.9.0

# Database & Cache
# ----------------
sqlalchemy==2.0.0
redis==5.0.0
motor==3.3.0  # Async MongoDB driver
alembic==1.12.0  # Database migrations

# Visualization
# -------------
plotly==5.17.0
matplotlib==3.7.0
seaborn==0.12.0

# AI/LLM Integration
# ------------------
anthropic==0.7.0
openai==1.3.0
langchain==0.0.340
transformers==4.30.0
sentence-transformers==2.2.2
nltk==3.8.1  # NLP basics

# Monitoring & Observability
# --------------------------
prometheus-client==0.18.0
grafana-api==1.0.3
sentry-sdk==1.38.0

# Authentication & Security
# -------------------------
pyjwt==2.8.0
cryptography==41.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async & Performance
# -------------------
asyncio==3.4.3
aioredis==2.0.1
celery==5.3.0
uvloop==0.19.0

# Utilities
# ---------
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.0
rich==13.6.0
tqdm==4.66.0
psutil==5.9.0

# Time Series Analysis
# --------------------
statsmodels==0.14.0
pmdarima==2.0.3
prophet==1.1.4
ruptures==1.1.8  # Change point detection
pykalman==0.9.7  # Kalman filtering
hmmlearn==0.3.0  # Hidden Markov Models

# Signal Processing
# -----------------
pywt==1.4.1  # Wavelets
networkx==3.1  # Network analysis
PyEMD==1.5.0  # Empirical Mode Decomposition
pyinform==0.2.0  # Information theory

# Deep Learning (Optional)
# ------------------------
# Uncomment if needed:
# tensorflow==2.13.0
# torch==2.0.0
# pytorch-forecasting==1.0.0

# Development & Testing
# ---------------------
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mypy==1.7.1
black==23.11.0
flake8==6.1.0
autopep8==2.0.4
pre-commit==3.5.0

# Documentation
# -------------
mkdocs==1.5.3
mkdocs-material==9.4.14
mkdocstrings[python]==0.24.0

# Deployment
# ----------
modal==0.55.4
docker==6.1.3
kubernetes==28.1.0
gunicorn==21.2.0

# Data Validation
# ---------------
pandera==0.17.2
great-expectations==0.18.0

# Additional Tools
# ----------------
beautifulsoup4==4.12.2  # Web scraping
Pillow==10.1.0  # Image processing
gymnasium==0.29.1  # RL environment

# Optional: GPU Support
# ---------------------
# Uncomment if using NVIDIA GPU:
# cuda-python==12.3.0
# cupy-cuda12x==12.3.0
# tensorrt==8.6.1
